refactor(extractor): moving files and cleaning code
