+++ src/scripts/content/extractor/domains/ArbitraryPage.js
