+++ src/scripts/content/extractor/domains/ExtractorPerplexity.spec.js
@@ -2,7 +2,7 @@
 // https://www.perplexity.ai/search/lorem-ipsum-dolor-g7PtZOxcTjuhYIMVTTYe.A
 // https://www.perplexity.ai/search/lorem-ipsum-dolor-2mBxUUVERqaFf8Q.pyYeYQ
 
-import ExtractorPerplexity from './ExtractorPerplexity';
+import ExtractorPerplexity from './Perplexity';
 import { formatLineBreaks } from "../../../shared/formatter/formatText";
 import { formatLink, initTurndown, setFileHeader, turndownConverter } from "../../../shared/formatter/formatMarkdown";
 import { sleep } from "../../../shared/utils/jsShorteners";
