+++ src/scripts/content/extractor/domains/MaxAIGoogle.js
