+++ src/scripts/content/extractor/domains/MaxAIGoogle.json
